refactor: move the SearchBar into component/ folder
