+++ src/features/search/components/SearchBar/SearchBar.type.ts
